requirements.txt
mysql-connector
mysql-connector-python-rf
mysql-connector-python
passlib
